ingester/requirements.txt
requests~=2.31.0
pydantic~=2.4.2
pydantic-settings~=2.0.3
#slack_sdk~=3.23.0
#taskw~=2.0.0
#gcsa~=2.1.0
click~=8.1.7
imap-tools~=1.4.0
atlassian-python-api~=3.41.2
markdownify~=0.11.6
beautifulsoup4~=4.12.2